requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
alembic
pydantic-settings
passlib[bcrypt]
python-dotenv
pip install python-jose[cryptography]
